Move TestHelper.race tests into @Isolated classes: batch 2 (#8075)
