+++ src/test/java/io/reactivex/rxjava4/internal/operators/observable/ObservableFlatMapTest.java
@@ -537,21 +537,6 @@ public class ObservableFlatMapTest extends RxJavaTest {
     }
 
     @Test
-    public void innerCompleteCancelRace() {
-        for (int i = 0; i < TestHelper.RACE_DEFAULT_LOOPS; i++) {
-            final PublishSubject<Integer> ps = PublishSubject.create();
-
-            final TestObserver<Integer> to = Observable.merge(Observable.just(ps)).test();
-
-            Runnable r1 = ps::onComplete;
-
-            Runnable r2 = to::dispose;
-
-            TestHelper.race(r1, r2);
-        }
-    }
-
-    @Test
     public void fusedInnerThrows() {
         Observable.just(1).hide()
         .flatMap((Function<Integer, ObservableSource<Object>>) _ -> Observable.range(1, 2).map(_ -> {
@@ -592,60 +577,6 @@ public class ObservableFlatMapTest extends RxJavaTest {
 
             assertTrue(list.contains("RxSi"), list.toString());
             assertTrue(list.contains("RxCo"), list.toString());
-        }
-    }
-
-    @Test
-    public void cancelScalarDrainRace() {
-        for (int i = 0; i < TestHelper.RACE_DEFAULT_LOOPS; i++) {
-            List<Throwable> errors = TestHelper.trackPluginErrors();
-            try {
-
-                final PublishSubject<Observable<Integer>> ps = PublishSubject.create();
-
-                final TestObserver<Integer> to = ps.flatMap(Functions.<Observable<Integer>>identity()).test();
-
-                Runnable r1 = to::dispose;
-                Runnable r2 = ps::onComplete;
-
-                TestHelper.race(r1, r2);
-
-                assertTrue(errors.isEmpty(), errors.toString());
-            } finally {
-                RxJavaPlugins.reset();
-            }
-        }
-    }
-
-    @Test
-    public void cancelDrainRace() {
-        for (int i = 0; i < TestHelper.RACE_DEFAULT_LOOPS; i++) {
-            for (int j = 1; j < 50; j += 5) {
-                List<Throwable> errors = TestHelper.trackPluginErrors();
-                try {
-
-                    final PublishSubject<Observable<Integer>> ps = PublishSubject.create();
-
-                    final TestObserver<Integer> to = ps.flatMap(Functions.<Observable<Integer>>identity()).test();
-
-                    final PublishSubject<Integer> just = PublishSubject.create();
-                    final PublishSubject<Integer> just2 = PublishSubject.create();
-                    ps.onNext(just);
-                    ps.onNext(just2);
-
-                    Runnable r1 = () -> {
-                        just2.onNext(1);
-                        to.dispose();
-                    };
-                    Runnable r2 = () -> just.onNext(1);
-
-                    TestHelper.race(r1, r2);
-
-                    assertTrue(errors.isEmpty(), errors.toString());
-                } finally {
-                    RxJavaPlugins.reset();
-                }
-            }
         }
     }
